requirements.txt
python-telegram==0.18.0
google-cloud-datastore==2.15.0
google-cloud-logging==3.5.0
